#1991 Split the module manager tests into two files, each one testing the newly split modules.
